Create a license file
This codebase does not have a license. I'd like developers to be able to extend this and potentially use it to host similar sites like let me bing that for you. Would it be possible for us to add a license, such as the MIT license?

The copyright line is based on this text from an earlier commit; feel free to change it if something else is more accurate:
Created by @coderifous, hosted by @rmm5t, inspired during lunch conversation with @coderifous, Tate, 
          @rmm5t, @EricStratton, and @methodvon. Feel free to suggest modifications to this license file based on that.